Rework interrupts to support phbs on higher chips
P8 IRQs are 19 bits, but we can only specify 16 bits in the IVTE
registers, so we were just throwing away the higher bits, which
correspond with the chip number.

Preserve the entire IRQ number throughout the driver, and only lop off
the top bits when actually writing to the IVTE_Offset, ErrIVTE and SERR
registers or corresponding fields in the process element.