Set AFU MMIO hang pulse to 4ms
Sapphire configures the PHB's MMIO timeout to ~16-32ms, so we need the
PSL to timeout AFU MMIOs before the PHB times it out.

A 4ms hang pulse means the PSL will time out MMIOs in ~4-8ms and the
next step up would be ~16-32ms which might race with the PHB, so ~4-8ms
is the largest we can use without changes in Sapphire.