Hook into PPC slb code to handle page size changes (not working)
On a 64k kernel there are situations where every page in a segment can
be degraded to 4k, which involves invalidating the SLBs, which we
weren't handling before.

One such situation involves mapping an IO region into userspace, which
we happen to do when we map in the problem state registers. Hence, we
started hitting this a lot all of a sudden when we switched to 64k
configs.

Sorry Alistair - I just broke modules again :-p