Fix wake up on forced detach; return other pending events before -EIO
This would have gone right back to sleep since none of pending_irq,
pending_fault or pending_afu_err were set - add !ctx->attached to that
list.

Do not return a bogus AFU error event in that case, since that event is
only for AFU errors from the PSL interrupt that set the AFU_Err
register - return an IO error instead.

Also, if there is a pending event and we have been detached, return the
pending event first, in case the pending event was related to the reason
we detached.