WIP fixing & cleaning afu/adapter init error paths
Looks like we could have double freed the error irq here and were not
unmapping the slice regs in the good case.

cxl_unregister_adapter has code that was originally in exit_capi and
should not be walking the adapter list.

Looking at these error paths I think we have a shitload of bugs waiting
for us - I need to spend some time cleaning up this mess and convincing
myself that it is correct.