Multiplex the PSL Interrupt
The PSL interrupt allows us to multiplex it between contexts as we can
tell which context it was for based on the PSL_PEHandle_An register
(alternatively the PSL_PID_TID_An register could be used).

Demultiplexing is suboptimal at the moment as it iterates over the
contexts list to find the matching context. We already use IPA to
allocate process handles - if we switch to IPR instead that will make
this a whole lot faster.

With the current aligned to power of 2 MSI interrupt allocator this
allows us to fit more contexts when the following number of AFU
interrupts are requested per context (not counting the PSL interrupt):
1, 2, 3, 4, 5, 8, 10, 16, 19, 21, 32, 35, 42, 64, 67, 128, 131, 256

Many cases end up as a tie because although we have more interrupts
available, multiplexing the PSL interrupt means we can only use 3
interrupt ranges for the AFU interrupts instead of 4.

There are also some cases (898 of them!) where we end up worse off.
These are generally higher number of interrupts per process and are less
likely to be used in practice - if this becomes an issue we can make the
multiplexing a module parameter or sysfs tunable, or change to a
different IRQ allocation scheme (e.g. use min_irqs_per_process search
size instead of the next power of 2). The cases where we hit this under
64 interrupts per process are: 11, 13, 22, 23, 25, 26, 27, 33, 43, 44,
45, 46, 47, 49, 50, 51, 52, 53, 54, 55

Of course, if different processes are using different numbers of
AFU interrupts all bets are off.