Remove unused context save area from start work ioctl + enforce reserved fields
Context save area is only for time sliced sharing mode, which we have no
support for yet - change those fields to reserved.

Also enforce reserved fields are 0 - thanks for the pointer mpe.

Also ensure that the argument passed to the check error ioctl is 0
incase we want to use that in the future.