s/BUG/WARN - Remove or replace BUGs with WARNs for upstreaming
The ones in the PSL error & Slice error interrupts are useful for
debugging, but we don't want to crash upstream kernels if there is a
chance we could recover.

We might want to consider disabling the SERR & PSL_Error IRQs to avoid
an IRQ storm from bad hardware like we have seen in the lab.

Most of the rest of these are just to guard against programming errors
- I removed a couple that are clearly impossible and changed the others
to warnings.