Remove FIR_RECOV; Add AFU_DEBUG; Use ErrStat
The addition of AFU_DEBUG in the latest PSL workbook made me notice that
FIR_RECOV has been removed as it used to occupy the same offset. I've
fixed up everywhere we used FIR_RECOV to use ErrStat instead.

AFU_DEBUG is exported via debugfs and printed out in the kernel log when
we hit various types of errors.