CXL: Fix inadvertently freeing PSL IRQ from MSI bitmap
The first range is reserved for the multiplexed PSL interrupt - fix the
code so it is not freed from the MSI bitmap when freeing AFU interrupts.

This bug meant that the PSL interrupt could be re-allocated as a user
AFU interrupt if the allocation code started scrounging for a single
interrupt (e.g. due to fragmentation or an AFU that only requires a
single interrupt), which would then cause it the virq to be freed
prematurely when that context completed and we would then blow up on the
next PSL interrupt.